feat: ajustes estruturais
