+++ TestingCRUD.sln
@@ -5,15 +5,15 @@ VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{B05B38FD-BC0D-41B2-A629-3CB7DB6D12E4}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestingCRUD.API", "src\TestingCRUD.API\TestingCRUD.API.csproj", "{800407DF-8B8B-4590-BA31-C7C2132A602D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestingCRUD.API", "src\TestingCRUD.API\TestingCRUD.API.csproj", "{800407DF-8B8B-4590-BA31-C7C2132A602D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestingCRUD.Infra", "src\TestingCRUD.Infra\TestingCRUD.Infra.csproj", "{129139E7-6FAD-4F37-BF8F-36FDE687C74D}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestingCRUD.Infra", "src\TestingCRUD.Infra\TestingCRUD.Infra.csproj", "{129139E7-6FAD-4F37-BF8F-36FDE687C74D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestingCRUD.Domain", "src\TestingCRUD.Domain\TestingCRUD.Domain.csproj", "{65A6195C-67F0-430B-A187-550B17098E63}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestingCRUD.Domain", "src\TestingCRUD.Domain\TestingCRUD.Domain.csproj", "{65A6195C-67F0-430B-A187-550B17098E63}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestingCRUD.Aplication", "src\TestingCRUD.Aplication\TestingCRUD.Aplication.csproj", "{39F38B94-A06A-4797-B6F6-79D9C6D05F43}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestingCRUD.Aplication", "src\TestingCRUD.Aplication\TestingCRUD.Aplication.csproj", "{39F38B94-A06A-4797-B6F6-79D9C6D05F43}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestingCRUD.Test", "src\TestingCRUD.Test\TestingCRUD.Test.csproj", "{7BF7C7BD-42C8-4DD6-82D6-44FBEB82DF19}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestingCRUD.Test", "src\TestingCRUD.Test\TestingCRUD.Test.csproj", "{C790AE44-B647-4D4A-81E4-635A23FD812D}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -37,10 +37,10 @@ Global
 		{39F38B94-A06A-4797-B6F6-79D9C6D05F43}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{39F38B94-A06A-4797-B6F6-79D9C6D05F43}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{39F38B94-A06A-4797-B6F6-79D9C6D05F43}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7BF7C7BD-42C8-4DD6-82D6-44FBEB82DF19}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7BF7C7BD-42C8-4DD6-82D6-44FBEB82DF19}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7BF7C7BD-42C8-4DD6-82D6-44FBEB82DF19}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7BF7C7BD-42C8-4DD6-82D6-44FBEB82DF19}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C790AE44-B647-4D4A-81E4-635A23FD812D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C790AE44-B647-4D4A-81E4-635A23FD812D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C790AE44-B647-4D4A-81E4-635A23FD812D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C790AE44-B647-4D4A-81E4-635A23FD812D}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -50,7 +50,7 @@ Global
 		{129139E7-6FAD-4F37-BF8F-36FDE687C74D} = {B05B38FD-BC0D-41B2-A629-3CB7DB6D12E4}
 		{65A6195C-67F0-430B-A187-550B17098E63} = {B05B38FD-BC0D-41B2-A629-3CB7DB6D12E4}
 		{39F38B94-A06A-4797-B6F6-79D9C6D05F43} = {B05B38FD-BC0D-41B2-A629-3CB7DB6D12E4}
-		{7BF7C7BD-42C8-4DD6-82D6-44FBEB82DF19} = {B05B38FD-BC0D-41B2-A629-3CB7DB6D12E4}
+		{C790AE44-B647-4D4A-81E4-635A23FD812D} = {B05B38FD-BC0D-41B2-A629-3CB7DB6D12E4}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {9445200C-003C-42FD-99C5-93A59DCE9345}
