+++ src/TestingCRUD.Aplication/Queries/CustomerQueries/GetCustomerByCpfQuery.cs
@@ -1,8 +1,8 @@
-﻿using TestingCRUD.Aplication.ViewModels;
+﻿using MediatR;
 
-using MediatR;
+using TestingCRUD.Aplication.ViewModels.CustomerViewModels;
 
-namespace TestingCRUD.Aplication.Queries
+namespace TestingCRUD.Aplication.Queries.CustomerQueries
 {
     public class GetCustomerByCpfQuery : IRequest<CustomerViewModel>
     {
