+++ src/TestingCRUD.Aplication/Queries/CustomerQueries/GetCustomersQuery.cs
@@ -1,0 +1,12 @@
+﻿using MediatR;
+
+using TestingCRUD.Aplication.ViewModels.CustomerViewModels;
+
+
+namespace TestingCRUD.Aplication.Queries.CustomerQueries
+{
+    public class GetCustomersQuery : IRequest<IEnumerable<CustomerViewModel>>
+    {
+
+    }
+}
