--- src/TestingCRUD.Aplication/Queries/GetCustomersQuery.cs
@@ -1,11 +1,0 @@
-﻿using TestingCRUD.Aplication.ViewModels;
-
-using MediatR;
-
-namespace TestingCRUD.Aplication.Queries
-{
-    public class GetCustomersQuery : IRequest<IEnumerable<CustomerViewModel>>
-    {
-
-    }
-}
