+++ src/TestingCRUD.Test/Class1.cs
@@ -1,7 +1,1 @@
-﻿namespace TestingCRUD.Test
-{
-    public class Class1
-    {
-
-    }
-}
+﻿
