+++ src/TestingCRUD.Test/UnitTest1.cs
@@ -1,0 +1,10 @@
+namespace TestingCRUD.Test;
+
+public class UnitTest1
+{
+    [Fact]
+    public void Test1()
+    {
+
+    }
+}
